Feat: `Footer` 추가 및 로그아웃 위치 변경
- 로그아웃 기능 진입 위치가 `SettingScreen` -> `HomePage`로 변경
- Footer를 통하여 팀 소개, 깃허브 페이지 이동, 로그아웃 기능 이용 가능하도록 추가
